requirements.txt.txt
makefile
asgiref==3.6.0
Django==5.0.4
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
mysqlclient==1.4.6